moved CPU state out of machine
