+++ SpiNNaker-comms/src/main/java/uk/ac/manchester/spinnaker/messages/scp/CountState.java
@@ -8,7 +8,7 @@ import static uk.ac.manchester.spinnaker.messages.sdp.SDPHeader.Flag.REPLY_EXPEC
 
 import java.nio.ByteBuffer;
 
-import uk.ac.manchester.spinnaker.machine.CPUState;
+import uk.ac.manchester.spinnaker.messages.model.CPUState;
 import uk.ac.manchester.spinnaker.messages.model.UnexpectedResponseCodeException;
 import uk.ac.manchester.spinnaker.messages.sdp.SDPHeader;
